refactor(frontend): central 作用域ページを FSD 化（_pages / widgets）
- _pages: central-dashboard、central-tenants（一覧・作成・編集）、central-settings
  の 3 slice。ページ実装は ui セグメント + index 公開 API
- widgets: sidebar / header（role 対応の単一実装のまま移設。central/store の
  分割は現実装が一体のため見送り — 分割は重複を生むだけで挙動同一）
- Next の app/central/**/page.tsx は re-export のみの薄殻に（URL 不変）
- features 抽出は見送り: 各ページのフォームはページ内で完結し再利用が無い。
  複数ページで共有される動作が現れた時点で抽出する
- Steiger: fsd/repetitive-naming を off（central-*/store-* の作用域接頭辞は
  D6 の設計判断で意図どおりのため）

Refs #199
