+++ frontend/src/_pages/central-dashboard/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardPage } from './ui/DashboardPage';
