+++ frontend/src/_pages/central-settings/index.ts
@@ -1,0 +1,1 @@
+export { default as SettingsPage } from './ui/SettingsPage';
