+++ frontend/src/_pages/central-tenants/index.ts
@@ -1,0 +1,3 @@
+export { default as TenantsPage } from './ui/TenantsPage';
+export { default as TenantCreatePage } from './ui/TenantCreatePage';
+export { default as TenantEditPage } from './ui/TenantEditPage';
