+++ frontend/src/app/central/tenants/page.tsx
@@ -1,348 +1,1 @@
-'use client';
-
-import { useEffect, useState, useCallback } from 'react';
-import { useAuth } from '@/contexts/AuthContext';
-import { useRouter } from 'next/navigation';
-import { Tenant, centralTenantApi } from '@/entities/tenant';
-import { PaginatedResponse } from '@/shared/api';
-import toast from 'react-hot-toast';
-
-export default function TenantsPage() {
-  const { logout } = useAuth();
-  const router = useRouter();
-  const [tenants, setTenants] = useState<PaginatedResponse<Tenant> | null>(null);
-  const [loadingTenants, setLoadingTenants] = useState(true);
-  const [searchTerm, setSearchTerm] = useState('');
-  const [currentPage, setCurrentPage] = useState(1);
-
-  const loadTenants = useCallback(async () => {
-    setLoadingTenants(true);
-    try {
-      const tenants = await centralTenantApi.getList({
-        page: currentPage,
-        per_page: 10,
-        search: searchTerm || undefined,
-      });
-
-      setTenants(tenants);
-    } catch (error) {
-      toast.error('店舗一覧の読み込みに失敗しました');
-    } finally {
-      setLoadingTenants(false);
-    }
-  }, [currentPage, searchTerm]);
-
-  useEffect(() => {
-    loadTenants();
-  }, [loadTenants, router]);
-
-  const handleSearch = (e: React.FormEvent) => {
-    e.preventDefault();
-    setCurrentPage(1);
-    loadTenants();
-  };
-
-  const handleDeleteTenant = async (id: string, name: string) => {
-    if (!confirm(`店舗「${name}」を削除しますか？この操作は取り消せません。`)) {
-      return;
-    }
-
-    try {
-      await centralTenantApi.delete(id);
-      toast.success('店舗を削除しました');
-      loadTenants();
-    } catch (error) {
-      toast.error('店舗の削除に失敗しました');
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-50">
-      {/* ナビゲーションバー */}
-      <nav className="bg-white shadow">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between h-16">
-            <div className="flex items-center space-x-4">
-              <button
-                onClick={() => router.push('/')}
-                className="text-indigo-600 hover:text-indigo-800 text-sm font-medium"
-              >
-                ← ダッシュボードへ戻る
-              </button>
-              <h1 className="text-xl font-semibold text-gray-900">店舗管理</h1>
-            </div>
-            <div className="flex items-center space-x-4">
-              <span className="text-sm text-gray-700">ようこそ、someone さん</span>
-              <button
-                onClick={logout}
-                className="text-gray-700 hover:text-gray-900 px-3 py-2 rounded-md text-sm font-medium"
-              >
-                ログアウト
-              </button>
-            </div>
-          </div>
-        </div>
-      </nav>
-
-      {/* 主要内容 */}
-      <div className="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
-        <div className="px-4 py-6 sm:px-0">
-          {/* ページヘッダー */}
-          <div className="md:flex md:items-center md:justify-between mb-6">
-            <div className="flex-1 min-w-0">
-              <h2 className="text-2xl font-bold leading-7 text-gray-900 sm:text-3xl sm:truncate">
-                店舗一覧
-              </h2>
-              <p className="mt-1 text-sm text-gray-500">システム内の全ての店舗を管理します</p>
-            </div>
-            <div className="mt-4 flex md:mt-0 md:ml-4">
-              <button
-                onClick={() => router.push('/central/tenants/create')}
-                className="ml-3 inline-flex items-center px-4 py-2 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-              >
-                <svg
-                  className="-ml-1 mr-2 h-5 w-5"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M12 6v6m0 0v6m0-6h6m-6 0H6"
-                  />
-                </svg>
-                店舗を追加
-              </button>
-            </div>
-          </div>
-
-          {/* 検索フォーム */}
-          <div className="bg-white shadow rounded-lg mb-6">
-            <div className="px-4 py-5 sm:p-6">
-              <form onSubmit={handleSearch} className="sm:flex sm:items-center">
-                <div className="w-full sm:max-w-xs">
-                  <label htmlFor="search" className="sr-only">
-                    店舗を検索
-                  </label>
-                  <input
-                    type="text"
-                    name="search"
-                    id="search"
-                    value={searchTerm}
-                    onChange={e => setSearchTerm(e.target.value)}
-                    className="shadow-sm focus:ring-indigo-500 focus:border-indigo-500 block w-full sm:text-sm border-gray-300 rounded-md"
-                    placeholder="店舗名またはドメインで検索..."
-                  />
-                </div>
-                <button
-                  type="submit"
-                  className="mt-3 w-full inline-flex items-center justify-center px-4 py-2 border border-transparent shadow-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
-                >
-                  検索
-                </button>
-                {searchTerm && (
-                  <button
-                    type="button"
-                    onClick={() => {
-                      setSearchTerm('');
-                      setCurrentPage(1);
-                    }}
-                    className="mt-3 w-full inline-flex items-center justify-center px-4 py-2 border border-gray-300 shadow-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
-                  >
-                    クリア
-                  </button>
-                )}
-              </form>
-            </div>
-          </div>
-
-          {/* 店舗一覧 */}
-          <div className="bg-white shadow overflow-hidden sm:rounded-md">
-            {loadingTenants ? (
-              <div className="px-4 py-12 text-center">
-                <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-indigo-600 mx-auto"></div>
-                <p className="mt-2 text-sm text-gray-500">読み込み中...</p>
-              </div>
-            ) : tenants && tenants.data.length > 0 ? (
-              <>
-                <ul className="divide-y divide-gray-200">
-                  {tenants.data.map(tenant => (
-                    <li key={tenant.id} className="px-4 py-4 hover:bg-gray-50">
-                      <div className="flex items-center justify-between">
-                        <div className="flex items-center min-w-0 flex-1">
-                          <div className="flex-shrink-0">
-                            <div className="h-12 w-12 rounded-full bg-indigo-100 flex items-center justify-center">
-                              <span className="text-lg font-medium text-indigo-600">
-                                {tenant.name.charAt(0).toUpperCase()}
-                              </span>
-                            </div>
-                          </div>
-                          <div className="ml-4 min-w-0 flex-1">
-                            <div className="flex items-center">
-                              <p className="text-lg font-medium text-gray-900 truncate">
-                                {tenant.name}
-                              </p>
-                              <span
-                                className={`ml-2 inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-                                  tenant.is_active
-                                    ? 'bg-green-100 text-green-800'
-                                    : 'bg-red-100 text-red-800'
-                                }`}
-                              >
-                                {tenant.is_active ? '有効' : '無効'}
-                              </span>
-                            </div>
-                          </div>
-                        </div>
-                        <div className="flex items-center space-x-2">
-                          <span className="text-sm text-gray-500">
-                            {new Date(tenant.created_at).toLocaleDateString('ja-JP')}
-                          </span>
-                          <div className="flex space-x-2">
-                            <button
-                              onClick={() => router.push(`/central/tenants/${tenant.id}/edit`)}
-                              className="inline-flex items-center px-3 py-1 border border-gray-300 shadow-sm text-sm leading-4 font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-                            >
-                              編集
-                            </button>
-                            <button
-                              onClick={() => handleDeleteTenant(tenant.id, tenant.name)}
-                              className="inline-flex items-center px-3 py-1 border border-transparent text-sm leading-4 font-medium rounded-md text-white bg-red-600 hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500"
-                            >
-                              削除
-                            </button>
-                          </div>
-                        </div>
-                      </div>
-                    </li>
-                  ))}
-                </ul>
-
-                {/* ページネーション */}
-                {tenants.last_page > 1 && (
-                  <div className="bg-white px-4 py-3 flex items-center justify-between border-t border-gray-200 sm:px-6">
-                    <div className="flex-1 flex justify-between sm:hidden">
-                      <button
-                        onClick={() => setCurrentPage(currentPage - 1)}
-                        disabled={currentPage <= 1}
-                        className="relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                      >
-                        前へ
-                      </button>
-                      <button
-                        onClick={() => setCurrentPage(currentPage + 1)}
-                        disabled={currentPage >= tenants.last_page}
-                        className="ml-3 relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                      >
-                        次へ
-                      </button>
-                    </div>
-                    <div className="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
-                      <div>
-                        <p className="text-sm text-gray-700">
-                          {tenants.total} 件中 {tenants.from}-{tenants.to} を表示
-                        </p>
-                      </div>
-                      <div>
-                        <nav className="relative z-0 inline-flex rounded-md shadow-sm -space-x-px">
-                          <button
-                            onClick={() => setCurrentPage(currentPage - 1)}
-                            disabled={currentPage <= 1}
-                            className="relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                          >
-                            <svg className="h-5 w-5" fill="currentColor" viewBox="0 0 20 20">
-                              <path
-                                fillRule="evenodd"
-                                d="M12.707 5.293a1 1 0 010 1.414L9.414 10l3.293 3.293a1 1 0 01-1.414 1.414l-4-4a1 1 0 010-1.414l4-4a1 1 0 011.414 0z"
-                                clipRule="evenodd"
-                              />
-                            </svg>
-                          </button>
-
-                          {/* ページ番号ボタン */}
-                          {Array.from({ length: Math.min(5, tenants.last_page) }, (_, i) => {
-                            const page = i + 1;
-                            return (
-                              <button
-                                key={page}
-                                onClick={() => setCurrentPage(page)}
-                                className={`relative inline-flex items-center px-4 py-2 border text-sm font-medium ${
-                                  page === currentPage
-                                    ? 'z-10 bg-indigo-50 border-indigo-500 text-indigo-600'
-                                    : 'bg-white border-gray-300 text-gray-500 hover:bg-gray-50'
-                                }`}
-                              >
-                                {page}
-                              </button>
-                            );
-                          })}
-
-                          <button
-                            onClick={() => setCurrentPage(currentPage + 1)}
-                            disabled={currentPage >= tenants.last_page}
-                            className="relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium text-gray-500 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
-                          >
-                            <svg className="h-5 w-5" fill="currentColor" viewBox="0 0 20 20">
-                              <path
-                                fillRule="evenodd"
-                                d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
-                                clipRule="evenodd"
-                              />
-                            </svg>
-                          </button>
-                        </nav>
-                      </div>
-                    </div>
-                  </div>
-                )}
-              </>
-            ) : (
-              <div className="px-4 py-12 text-center">
-                <svg
-                  className="mx-auto h-12 w-12 text-gray-400"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M19 21V5a2 2 0 00-2-2H7a2 2 0 00-2 2v16m14 0h2m-2 0h-5m-9 0H3m2 0h5M9 7h1m-1 4h1m4-4h1m-1 4h1m-5 10v-5a1 1 0 011-1h2a1 1 0 011 1v5m-4 0h4"
-                  />
-                </svg>
-                <h3 className="mt-2 text-sm font-medium text-gray-900">店舗がありません</h3>
-                <p className="mt-1 text-sm text-gray-500">
-                  {searchTerm ? '該当する店舗が見つかりません' : '最初の店舗を作成しましょう'}
-                </p>
-                <div className="mt-6">
-                  <button
-                    onClick={() => router.push('/central/tenants/create')}
-                    className="inline-flex items-center px-4 py-2 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-                  >
-                    <svg
-                      className="-ml-1 mr-2 h-5 w-5"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M12 6v6m0 0v6m0-6h6m-6 0H6"
-                      />
-                    </svg>
-                    店舗を追加
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
+export { TenantsPage as default } from '@/_pages/central-tenants';
